gecko/CMakeLists: don't error out on missing SOC_GECKO_SERIES1/2
Right now, CMake errors out if either `CONFIG_SOC_GECKO_SERIES1` or
`CONFIG_SOC_GECKO_SERIES2` is undefined. This check is for adding the
right RAIL headers for the chosen SoC series.

Using RAIL is not required for normal SoC operation in Zephyr, so
erroring out here is not required.

Signed-off-by: Filip Kokosinski <fkokosinski@antmicro.com>