zephyr: Add allocation API shim for SiSDK memory manager
Simplicity SDK drivers and services make use of the memory manager
API to perform dynamic memory management. Redirect calls to the
basic allocation functions to the standard library, which is
backed by a sys_heap implementation in Zephyr.

This leaves most of the memory manager API unimplemented, but
drivers depending on it are not yet used in Zephyr, so we can
get away with it. A long-term solution for the two memory
management systems to coexist will be needed in the future.

Signed-off-by: Aksel Skauge Mellbye <aksel.mellbye@silabs.com>