wiseconnect: Avoid compiler warning in PRINT_STATUS()
uint32_t is sometime defined as an unsigned int and sometime as a long
unsigned int. So, it is not possible to display it with %lu or with %u.

The portable way to display uint32_t is to use %"PRIu32" provided by
inttypes.h.

An alternative is to cast the uint32_t to explicitly set the proper
length.

Upstream-status: Pending
Signed-off-by: Jérôme Pouiller <jerome.pouiller@silabs.com>