wiseconnect: Do not use .common_ipmu_ram
Wiseconnect declare the IPMU data in .common_ipmu_ram section. This
section is then locate in RAM or in PSRAM depending of the
configuration. It is not clear if these data are in RAM for performance
purpose or if they are acceded during critical section when the Flash
is not available.

Anyway, these structures are not marked with "const" nor "static", so
Zephyr would place them in RAM with the rest of the data. This patch
explicitly places them in .data, just in case.

Note this reverts the change introduced on Wiseconnect 3.4.

Upstream-status: Pending
Signed-off-by: Jérôme Pouiller <jerome.pouiller@silabs.com>