scripts: add script to generate the comparator silabs-acmp bindings header
The script added will combine all possible ACMP positive and negative
input values and define helper macros to make it easier for applications
to select ACMP inputs for their comparator driver. The script also
verifies that the integer input values are consistent between the parts
that this header is generated for. However, it must be noted that some
of the input defines may not be available as positive and/or negative
inputs for some devices. It's recommended to always reference your
part's design book before selecting any of the inputs in this generated
header.

Signed-off-by: Christian Galante <christian.galante@silabs.com>