wiseconnect: Get rid of PSRAM_BASE_ADDRESS
PSRAM_BASE_ADDRESS contains an hardware configuration parameter. This
parameter belongs to the DT on the Zephyr side. Hence, it can't be
defined with a macro.

Fortunately, this parameter can be deducted from the Chip Select number.
So, this patch introduce a getter to retrieve this value. This is also
the opportunity to factorize the code that check the access boundaries.

Signed-off-by: Jérôme Pouiller <jerome.pouiller@silabs.com>