wiseconnect: Resolved issues when multiple events are set in si91x bus events.
It was observed that when one event is set and another event is subsequently
set, the first event gets cleared prematurely before it can be processed. To
address this, the osEventFlagsWait call was updated to use the osFlagsNoClear
option, which prevents automatic clearing of events upon retrieval. Events
are now manually cleared after they are read, ensuring that all set events
are reliably processed and none are missed due to premature clearing caused
by subsequent event updates.

Signed-off-by: Rahul Gurram <rahul.gurram@silabs.com>