wiseconnect: Split PSRAM init()/uninit() functions
sl_si91x_psram_init() and sl_si91x_psram_uninit() are also in charge of
configuring the pinctrl and the clocks. A workaround have been
introduced to avoid change in pinctrl but they still changed the clock
configuration.

We definitely need to expose the clock configuration to Zephyr users.
Thus, this patch split the sl_si91x_psram_*init() functions in smaller
pieces. So it is possible to configure the device without changing the
clock or the pinctrl.

Following this patch, we may drop:
  9cdff97f "wiseconnect: Leave Zephyr pinctrl managing the pin-muxing"
  ddb5a104 "wiseconnect: Remove dependency on egpio_set_*"

Upstream-Status: Pending
Signed-off-by: Jérôme Pouiller <jerome.pouiller@silabs.com>